feat: separate autobump threads module
