+++ src/commands/autobump-threads/remove-thread.test.ts
